Add support for Server Name Indication. Done during IETF 103 hackathon
This patch adds support for SNI (Server Name Indication). The option
--tls-server-name has been added in order to specify the server name
to be used in the SeverName extension sent in the ClientHello for TLS
connections. The newly added --tls-server-name option is optional:
If not specified, httperf will simply function as usual by omitting
the ServerName extension in the ClientHello for TLS connections.

Closes: #47
Signed-off-by: Muzaffar Auhammud <muzaffar@cyberstorm.mu>